requirements-models.txt
numpy
pandas
matplotlib
scikit-learn
seaborn
torch
tensorflow
transformers
pgmpy
weaviate-client>=3.26.7,<4.0.0
networkx
requests
streamlit>=1.35.0
absl-py==2.4.0
annotated-doc==0.0.4
anyio==4.12.1
astunparse==1.6.3
Authlib==1.6.9
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
filelock==3.25.1
flatbuffers==25.12.19
fonttools==4.62.0
fsspec==2026.2.0
gast==0.7.0
git-filter-repo==2.47.0
google-pasta==0.2.0
grpcio==1.78.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.4.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.6.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
keras==3.13.2
kiwisolver==1.5.0
libclang==18.1.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
namex==0.1.0
networkx==3.6.1
numpy==2.4.3
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pandas==3.0.1
patsy==1.0.2
pgmpy==1.0.0
pillow==12.1.1
protobuf>=6.31.1,<7
pycparser==3.0
Pygments==2.19.2
pyparsing==3.3.2
pyro-api==0.1.2
pyro-ppl==1.9.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
statsmodels==0.14.6
sympy==1.14.0
tensorflow==2.21.0
termcolor==3.3.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
validators==0.35.0
weaviate-client==3.26.7
wheel==0.46.3
wrapt==2.1.2
